Added proper timing
* For sphere rotation
* For observatory positioning

Note: the original code seems to have a bug which explains a slight discrepancy between this new and the original visualisation.
In the original code gmt_dtime() is used to get the GMT time, but this function is a mere #define of dtime() under unix, hence any
difference of dtime() and gmt_dtime() will result in 0 instead of the real UTC/GMT offset.

Note: dtime() returns local time...